+++ packages/std/src/exports.rs
@@ -15,10 +15,10 @@ use serde::de::DeserializeOwned;
 use crate::deps::OwnedDeps;
 #[cfg(feature = "stargate")]
 use crate::ibc::{
-    IbcBasicResponse, IbcChannelCloseMsg, IbcChannelConnectMsg, IbcChannelOpenMsg,
-    IbcChannelOpenResponse, IbcPacketAckMsg, IbcPacketReceiveMsg, IbcPacketTimeoutMsg,
-    IbcReceiveResponse,
+    IbcBasicResponse, IbcChannelCloseMsg, IbcChannelConnectMsg, IbcPacketAckMsg,
+    IbcPacketReceiveMsg, IbcPacketTimeoutMsg, IbcReceiveResponse,
 };
+use crate::ibc::{IbcChannelOpenMsg, IbcChannelOpenResponse};
 use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 use crate::memory::{alloc, consume_region, release_buffer, Region};
 #[cfg(feature = "abort")]
@@ -529,7 +529,6 @@ where
     query_fn(deps.as_ref(), env, msg).into()
 }
 
-#[cfg(feature = "stargate")]
 fn _do_ibc_channel_open<Q, E>(
     contract_fn: &dyn Fn(DepsMut<Q>, Env, IbcChannelOpenMsg) -> Result<IbcChannelOpenResponse, E>,
     env_ptr: *mut Region,