+++ packages/std/src/ibc.rs
@@ -5,7 +5,6 @@ use core::cmp::{Ord, Ordering, PartialOrd};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-#[cfg(feature = "ibc3")]
 use crate::addresses::Addr;
 use crate::binary::Binary;
 use crate::coin::Coin;
@@ -299,11 +298,9 @@ impl From<IbcChannelOpenMsg> for IbcChannel {
     }
 }
 
-/// Note that this serializes as "null".
-#[cfg(not(feature = "ibc3"))]
-pub type IbcChannelOpenResponse = ();
-/// This serializes either as "null" or a JSON object.
-#[cfg(feature = "ibc3")]
+/// This serializes either as `null` or a JSON object.
+/// Within the response, a channel version can be specified.
+/// If `null` is provided instead, the incoming channel version is accepted.
 pub type IbcChannelOpenResponse = Option<Ibc3ChannelOpenResponse>;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
@@ -405,17 +402,10 @@ impl From<IbcChannelCloseMsg> for IbcChannel {
 #[non_exhaustive]
 pub struct IbcPacketReceiveMsg {
     pub packet: IbcPacket,
-    #[cfg(feature = "ibc3")]
     pub relayer: Addr,
 }
 
 impl IbcPacketReceiveMsg {
-    #[cfg(not(feature = "ibc3"))]
-    pub fn new(packet: IbcPacket) -> Self {
-        Self { packet }
-    }
-
-    #[cfg(feature = "ibc3")]
     pub fn new(packet: IbcPacket, relayer: Addr) -> Self {
         Self { packet, relayer }
     }
@@ -427,20 +417,10 @@ impl IbcPacketReceiveMsg {
 pub struct IbcPacketAckMsg {
     pub acknowledgement: IbcAcknowledgement,
     pub original_packet: IbcPacket,
-    #[cfg(feature = "ibc3")]
     pub relayer: Addr,
 }
 
 impl IbcPacketAckMsg {
-    #[cfg(not(feature = "ibc3"))]
-    pub fn new(acknowledgement: IbcAcknowledgement, original_packet: IbcPacket) -> Self {
-        Self {
-            acknowledgement,
-            original_packet,
-        }
-    }
-
-    #[cfg(feature = "ibc3")]
     pub fn new(
         acknowledgement: IbcAcknowledgement,
         original_packet: IbcPacket,
@@ -459,17 +439,10 @@ impl IbcPacketAckMsg {
 #[non_exhaustive]
 pub struct IbcPacketTimeoutMsg {
     pub packet: IbcPacket,
-    #[cfg(feature = "ibc3")]
     pub relayer: Addr,
 }
 
 impl IbcPacketTimeoutMsg {
-    #[cfg(not(feature = "ibc3"))]
-    pub fn new(packet: IbcPacket) -> Self {
-        Self { packet }
-    }
-
-    #[cfg(feature = "ibc3")]
     pub fn new(packet: IbcPacket, relayer: Addr) -> Self {
         Self { packet, relayer }
     }