+++ packages/vm/src/calls.rs
@@ -825,9 +825,10 @@ mod tests {
         use crate::testing::{
             mock_env, mock_info, mock_instance, MockApi, MockQuerier, MockStorage,
         };
+        use cosmwasm_std::testing::mock_ibc_packet_timeout;
         use cosmwasm_std::testing::{
             mock_ibc_channel_close_init, mock_ibc_channel_connect_ack, mock_ibc_channel_open_init,
-            mock_ibc_packet_ack, mock_ibc_packet_recv, mock_ibc_packet_timeout, mock_wasmd_attr,
+            mock_ibc_packet_ack, mock_ibc_packet_recv, mock_wasmd_attr,
         };
         use cosmwasm_std::{
             Empty, Event, IbcAcknowledgement, IbcOrder, Reply, ReplyOn, SubMsgResponse,